Fixed SBTC_ERRNOPTR macro to work also on x86-64 (errno is int on x86-64 and sizeof(int) == sizeof(long) on i386, so there's no need for another comparison).
git-svn-id: https://svn.aros.org/svn/aros/trunk@28701 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ contrib/Networking/Stacks/AROSTCP/netinclude/bsdsocket/socketbasetags.h
@@ -119,7 +119,7 @@
  * only 1,2 & 4 are legal 'size' values. If the 'size' value is illegal,
  * the tag is set to 0, which causes SocketBaseTagList() to fail.
  */
-#define SBTC_ERRNOPTR(size)    ((size == sizeof(long)) ? SBTC_ERRNOLONGPTR   :\
+#define SBTC_ERRNOPTR(size)    ((size == sizeof(int)) ? SBTC_ERRNOLONGPTR   :\
 				((size == sizeof(short)) ? SBTC_ERRNOWORDPTR :\
 				 ((size == sizeof(char)) ? SBTC_ERRNOBYTEPTR :\
 				  0)))
